Fix bookmark feature
Before, when adding a folder of bookmarks, not all bookmarks
weren't added to the dump due to the local storage interractions.
There was many concurrent calls to the local storage and they all
wrote what they know at the time they were accessing it. Thus we
add data loss.
Now the process is rewritten to make a single call with all links
needed to be dumped.