Change clipboard copy action.
Before, the copy action was slow because it was playing in the DOM
to be able to copy the select content to the clipboard.
Now, the copy action use the modern API to copy to clipboard. There
is no need to play with the DOM anymore. Thus making the whole
process a lot simpler.

See #34