Change variable names in content scripts
Before, it was not possible to call the scrape content script then
the capture content script because their variables shared their
name.
Now, the names are differents allowing to call the 2 scripts.

There is still a problem when calling the same script multiple times.
The reason is identical. It's a known limitation at the moment. I will
probably fix that in the future but for now, the solution is to refresh
the page before calling the same content script.